pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forklone"
version = "0.1.0"
description = "Fork & clone a GitHub repository"
readme = "README.rst"
requires-python = ">=3.7"
license = "MIT"
license-files = { paths = ["LICENSE"] }
authors = [
    { name = "John Thorvald Wodder II", email = "forklone@varonathe.org" }
]

###keywords = [ ]

classifiers = [
    "Private :: Keep Off PyPI",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]

dependencies = [
    "click >= 8.0",
    "ghrepo ~= 0.1",
    "ghtoken ~= 0.1",
    "PyGithub ~= 2.0",
]

[project.scripts]
forklone = "forklone:main"

[project.urls]
"Source Code" = "https://github.com/jwodder/forklone"
"Bug Tracker" = "https://github.com/jwodder/forklone/issues"

[tool.hatch.version]
path = "forklone.py"

[tool.hatch.envs.default]
python = "3"
